Switch to Input System for Android
